+++ src/test/java/com/bitmart/unit/api/TestAccount.java
@@ -1,10 +1,10 @@
-package com.bitmart.api;
+package com.bitmart.unit.api;
 
 import com.bitmart.api.common.CloudException;
 import com.bitmart.api.common.CloudResponse;
 import com.bitmart.api.request.account.prv.*;
 import com.bitmart.api.request.account.pub.AccountCurrenciesRequest;
-import com.bitmart.data.TestData;
+import com.bitmart.unit.data.TestData;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;